working fix

+++ Sample_Bugs/AdvancedAverageBug.java
@@ -2,7 +2,7 @@ public class AdvancedAverageBug{
     public static void main(String[] args) {
         int total = 7;
         int count = 2;
-        double average = (double) total / count;
+        double average = total / count;
         System.out.println("Average is: " + average);
     }
 }
